+++ app/assets/javascripts/application.js
@@ -69,4 +69,4 @@
 //= require admin/exports
 //= require admin/engagement_heatmaps
 //= require admin/research_articles
-//= require social/sleep_tips
+//= require admin/social_media