+++ app/controllers/admin_controller.rb
@@ -9,6 +9,9 @@ class AdminController < ApplicationController
   def dashboard
   end
 
+  def social_media
+  end
+
   def research_topics
     @research_topics = ResearchTopic.where.not(topic_id: nil).order("created_at desc")
   end