+++ app/controllers/static_controller.rb
@@ -4,8 +4,6 @@ class StaticController < ApplicationController
   before_action :load_pc, only: [ :team, :advisory, :learn, :research ]
   before_action :about_layout, only: :research
 
-  before_action :check_owner_or_moderator, only: :sleep_tips
-
   before_action :set_SEO_elements
 
   layout 'static'
@@ -145,11 +143,6 @@ class StaticController < ApplicationController
 
   def side_effects_PAP
     render 'static/SEO_content/side_effects_PAP'
-  end
-
-  # Social media content
-  def sleep_tips
-    render 'static/sleep_tips/sleep_tips'
   end
 
   ## THEME